# perf(feistel): use blake2b_simd