--- a/storage-proofs/Cargo.toml
+++ b/storage-proofs/Cargo.toml
@@ -36,6 +36,7 @@ slog = { version = "2.4.1", features = ["max_level_trace", "release_max_level_tr
 serde = "1.0"
 serde_derive = "1.0"
 base64 = "0.10.0"
+blake2b_simd = "0.4.1"
 
 [dependencies.pairing]
 version = "0.14.2"