before

```
replication_time: 4.609915269s, target: stats, place: filecoin-proofs/examples/zigzag.rs:193 zigzag, root: filecoin-proofs
replication_time/byte: 4.395µs, target: stats, place: filecoin-proofs/examples/zigzag.rs:194 zigzag, root: filecoin-proofs
```

after
```
replication_time: 3.887283333s, target: stats, place: filecoin-proofs/examples/zigzag.rs:193 zigzag, root: filecoin-proofs
replication_time/byte: 3.707µs, target: stats, place: filecoin-proofs/examples/zigzag.rs:194 zigzag, root: filecoin-proofs
```